refactor: semantic chunker
