+++ jchunk-core/src/main/java/jchunk/chunker/core/decorators/VisibleForTesting.java
@@ -1,0 +1,16 @@
+package jchunk.chunker.core.decorators;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Annotation to indicate that the visibility of a method, field, or class is increased
+ * for testing purposes.
+ */
+@Retention(RetentionPolicy.CLASS)
+@Target({ ElementType.METHOD, ElementType.FIELD, ElementType.CONSTRUCTOR, ElementType.TYPE })
+public @interface VisibleForTesting {
+
+}
