+++ jchunk-semantic/src/test/java/jchunk/chunker/semantic/SemanticChunkerIT.java
@@ -1,6 +1,7 @@
 package jchunk.chunker.semantic;
 
 import jchunk.chunker.core.chunk.Chunk;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.ai.embedding.EmbeddingModel;
@@ -14,6 +15,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.core.io.DefaultResourceLoader;
 
 import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 
@@ -22,6 +25,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest
 @Disabled("Only for manual testing purposes.")
 class SemanticChunkerIT {
+
+	private static Method splitSentences;
+
+	private static Method combineSentences;
+
+	private static Method embedSentences;
+
+	private static Method calculateSimilarities;
 
 	@Autowired
 	private SemanticChunker semanticChunker;
@@ -41,21 +52,48 @@ class SemanticChunkerIT {
 		}
 	}
 
+	static void initPrivateMethods() throws NoSuchMethodException {
+		splitSentences = SemanticChunker.class.getDeclaredMethod("splitSentences", String.class,
+				SentenceSplitingStrategy.class);
+		combineSentences = SemanticChunker.class.getDeclaredMethod("combineSentences", List.class, Integer.class);
+		embedSentences = SemanticChunker.class.getDeclaredMethod("embedSentences", EmbeddingModel.class, List.class);
+		calculateSimilarities = SemanticChunker.class.getDeclaredMethod("calculateSimilarities", List.class);
+		splitSentences.setAccessible(true);
+		combineSentences.setAccessible(true);
+		embedSentences.setAccessible(true);
+		calculateSimilarities.setAccessible(true);
+	}
+
+	@BeforeAll
+	static void init() throws NoSuchMethodException {
+		initPrivateMethods();
+	}
+
 	@Test
 	void documentContentLoaded() {
 		assertThat(mitContent).isNotBlank();
 	}
 
 	@Test
-	void getSentences() {
-		List<Sentence> sentences = Utils.splitSentences(mitContent, SentenceSplitingStrategy.DEFAULT);
+	void getChunks() {
+		List<Chunk> chunks = this.semanticChunker.split(mitContent);
+		assertThat(chunks).isNotEmpty();
+	}
+
+	@Test
+	@SuppressWarnings("unchecked")
+	void getSentences() throws InvocationTargetException, IllegalAccessException {
+		List<Sentence> sentences = (List<Sentence>) splitSentences.invoke(semanticChunker, mitContent,
+				SentenceSplitingStrategy.DEFAULT);
 		assertThat(sentences).isNotEmpty().hasSize(317);
 	}
 
 	@Test
-	void combineSentences() {
-		List<Sentence> sentences = Utils.splitSentences(mitContent, SentenceSplitingStrategy.DEFAULT);
-		List<Sentence> combined = Utils.combineSentences(sentences, 1);
+	@SuppressWarnings("unchecked")
+	void combineSentences() throws InvocationTargetException, IllegalAccessException {
+		List<Sentence> sentences = (List<Sentence>) splitSentences.invoke(semanticChunker, mitContent,
+				SentenceSplitingStrategy.DEFAULT);
+		List<Sentence> combined = (List<Sentence>) combineSentences.invoke(semanticChunker, sentences, 1);
 
 		assertThat(combined).isNotEmpty();
 		assertThat(combined).hasSize(317);
@@ -67,12 +105,15 @@ class SemanticChunkerIT {
 	}
 
 	@Test
-	void embedChunks() {
+	@SuppressWarnings("unchecked")
+	void embedChunks() throws InvocationTargetException, IllegalAccessException {
+
 		int EMBEDDING_MODEL_DIMENSION = 384;
 
-		List<Sentence> sentences = Utils.splitSentences(mitContent, SentenceSplitingStrategy.DEFAULT);
-		List<Sentence> combined = Utils.combineSentences(sentences, 1);
-		List<Sentence> embedded = Utils.embedSentences(embeddingModel, combined);
+		List<Sentence> sentences = (List<Sentence>) splitSentences.invoke(semanticChunker, mitContent,
+				SentenceSplitingStrategy.DEFAULT);
+		List<Sentence> combined = (List<Sentence>) combineSentences.invoke(semanticChunker, sentences, 1);
+		List<Sentence> embedded = (List<Sentence>) embedSentences.invoke(semanticChunker, embeddingModel, combined);
 
 		assertThat(embedded).isNotEmpty().hasSize(317);
 
@@ -84,19 +125,15 @@ class SemanticChunkerIT {
 	}
 
 	@Test
-	void getCosineDistancesArray() {
-		List<Sentence> sentences = Utils.splitSentences(mitContent, SentenceSplitingStrategy.DEFAULT);
-		List<Sentence> combined = Utils.combineSentences(sentences, 1);
-		List<Sentence> embedded = Utils.embedSentences(embeddingModel, combined);
-		List<Double> distances = Utils.calculateSimilarities(embedded);
+	@SuppressWarnings("unchecked")
+	void getCosineDistancesArray() throws InvocationTargetException, IllegalAccessException {
+		List<Sentence> sentences = (List<Sentence>) splitSentences.invoke(semanticChunker, mitContent,
+				SentenceSplitingStrategy.DEFAULT);
+		List<Sentence> combined = (List<Sentence>) combineSentences.invoke(semanticChunker, sentences, 1);
+		List<Sentence> embedded = (List<Sentence>) embedSentences.invoke(semanticChunker, embeddingModel, combined);
+		List<Double> distances = (List<Double>) calculateSimilarities.invoke(semanticChunker, embedded);
 
 		assertThat(distances).hasSize(sentences.size() - 1);
-	}
-
-	@Test
-	void getChunks() {
-		List<Chunk> chunks = this.semanticChunker.split(mitContent);
-		assertThat(chunks).isNotEmpty();
 	}
 
 	@SpringBootConfiguration
